Move domain::Tags::ObjectCenter to DomainCreator library
