+++ src/ControlSystem/ControlErrors/Rotation.hpp
@@ -12,8 +12,8 @@
 #include "DataStructures/DataBox/DataBox.hpp"
 #include "DataStructures/DataVector.hpp"
 #include "DataStructures/Tensor/Tensor.hpp"
+#include "Domain/Creators/Tags/ObjectCenter.hpp"
 #include "Domain/Structure/ObjectLabel.hpp"
-#include "Domain/Tags/ObjectCenter.hpp"
 #include "Options/Options.hpp"
 #include "Parallel/GlobalCache.hpp"
 #include "Utilities/ErrorHandling/Assert.hpp"
