+++ src/Domain/Tags/CMakeLists.txt
@@ -1,12 +1,6 @@
 # Distributed under the MIT License.
 # See LICENSE.txt for details.
 
-spectre_target_sources(
-  ${LIBRARY}
-  PRIVATE
-  ObjectCenter.cpp
-  )
-
 spectre_target_headers(
   ${LIBRARY}
   INCLUDE_DIRECTORY ${CMAKE_SOURCE_DIR}/src
@@ -14,6 +8,5 @@ spectre_target_headers(
   ExternalBoundaryConditions.hpp
   FaceNormal.hpp
   Faces.hpp
-  ObjectCenter.hpp
   SurfaceJacobian.hpp
   )
